SCISOFT-CORE: add tag to remove compilation error from generated code

+++ uk.ac.diamond.scisoft.analysis/src/uk/ac/diamond/scisoft/analysis/dataset/CompoundFloatDataset.java
@@ -951,7 +951,7 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 		IndexIterator iter = getIterator(); // REAL_ONLY
 		while (iter.hasNext()) { // REAL_ONLY
 			for (int i = 0; i < isize; i++) { // REAL_ONLY
-				double x = data[iter.index + i]; // REAL_ONLY
+				float x = data[iter.index + i]; // PRIM_TYPE // REAL_ONLY
 				if (Float.isNaN(x) || Float.isInfinite(x)) // CLASS_TYPE // REAL_ONLY
 					return true; // REAL_ONLY
 			} // REAL_ONLY
